split grid components
