+++ apps/forms/scaffolds/grid-editor/index.tsx
@@ -15,63 +15,19 @@ import {
 import toast from "react-hot-toast";
 import { useEditorState } from "../editor";
 import Link from "next/link";
-import {
-  CommitIcon,
-  Cross2Icon,
-  DotIcon,
-  DownloadIcon,
-  GearIcon,
-  PieChartIcon,
-  PlusIcon,
-  ReloadIcon,
-  TrashIcon,
-} from "@radix-ui/react-icons";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
+import { DownloadIcon, PieChartIcon, TrashIcon } from "@radix-ui/react-icons";
 import { Badge } from "@/components/ui/badge";
-import {
-  DropdownMenu,
-  DropdownMenuCheckboxItem,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuRadioGroup,
-  DropdownMenuRadioItem,
-  DropdownMenuSeparator,
-  DropdownMenuShortcut,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
 import { Button, buttonVariants } from "@/components/ui/button";
-import { format, startOfDay, addSeconds } from "date-fns";
-import { format as formatTZ } from "date-fns-tz";
-import { LOCALTZ, tztostr } from "../editor/symbols";
 import { GridData } from "./grid-data";
 import clsx from "clsx";
 import { Tabs, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import type { XSupabaseQuery } from "@/lib/supabase-postgrest/builder";
-import { Spinner } from "@/components/spinner";
-import { SearchInput } from "@/components/extension/search-input";
-import { useDebounceCallback } from "usehooks-ts";
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipTrigger,
-} from "@/components/ui/tooltip";
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
-import { ArrowDownUpIcon } from "lucide-react";
-import { cn } from "@/utils";
-import { PopoverClose } from "@radix-ui/react-popover";
-import { FormFieldTypeIcon } from "@/components/form-field-type-icon";
 import { SupabaseLogo } from "@/components/logos";
+import { MaxRowsSelect } from "./components/limit";
+import { GridViewSettings } from "./components/view-settings";
+import { DataLoadingIndicator } from "./components/refresh";
+import { XSupaDataGridSort } from "./components/sort";
+import { DataGridLocalSearch } from "./components/search";
 
 export function GridEditor() {
   const [state, dispatch] = useEditorState();
@@ -318,212 +274,6 @@ function TableTools() {
   );
 }
 
-/**
- * this can also be used for form query, but at this moment, form does not have a db level field sorting query.
- * plus, this uses the column name, which in the future, it should be using field id for more universal handling.
- * when it updates to id, the x-supabase query route will also have to change.
- */
-function XSupaDataGridSort() {
-  const [state, dispatch] = useEditorState();
-
-  const { fields, datagrid_orderby } = state;
-
-  const isset = Object.keys(datagrid_orderby).length > 0;
-
-  const onReset = useCallback(() => {
-    dispatch({ type: "editor/data-grid/orderby/reset" });
-  }, [dispatch]);
-
-  const onAdd = useCallback(
-    (column_id: string) => {
-      dispatch({
-        type: "editor/data-grid/orderby",
-        column_id: column_id,
-        data: {},
-      });
-    },
-    [dispatch]
-  );
-
-  const onUpdate = useCallback(
-    (column_id: string, data: { ascending?: boolean }) => {
-      dispatch({
-        type: "editor/data-grid/orderby",
-        column_id: column_id,
-        data: data,
-      });
-    },
-    [dispatch]
-  );
-
-  const onRemove = useCallback(
-    (column_id: string) => {
-      dispatch({
-        type: "editor/data-grid/orderby",
-        column_id: column_id,
-        data: null,
-      });
-    },
-    [dispatch]
-  );
-
-  return (
-    <Popover modal>
-      <PopoverTrigger asChild>
-        <Button
-          variant="ghost"
-          size="icon"
-          className={cn(
-            "relative",
-            "text-muted-foreground",
-            isset && " text-accent-foreground"
-          )}
-        >
-          <ArrowDownUpIcon className="w-4 h-4" />
-          {isset && <DotIcon className="absolute top-0.5 right-0.5" />}
-        </Button>
-      </PopoverTrigger>
-      <PopoverContent className="p-2 max-w-none">
-        <section className="py-2">
-          <div className="flex flex-col space-y-2">
-            {Object.keys(datagrid_orderby).map((col) => {
-              const orderby = datagrid_orderby[col];
-              return (
-                <div key={col} className="flex gap-2 px-2">
-                  <div className="flex items-center gap-2 flex-1">
-                    <div>
-                      <Select
-                        disabled
-                        value={orderby.column}
-                        onValueChange={(value) => {
-                          onUpdate(value, orderby);
-                        }}
-                      >
-                        <SelectTrigger>
-                          <SelectValue />
-                        </SelectTrigger>
-                        <SelectContent>
-                          {fields.map((field) => (
-                            <SelectItem key={field.name} value={field.name}>
-                              {field.name}
-                            </SelectItem>
-                          ))}
-                        </SelectContent>
-                      </Select>
-                    </div>
-                    <div>
-                      <Select
-                        value={orderby.ascending ? "ASC" : "DESC"}
-                        onValueChange={(value) => {
-                          onUpdate(orderby.column, {
-                            ascending: value === "ASC",
-                          });
-                        }}
-                      >
-                        <SelectTrigger>
-                          <SelectValue />
-                        </SelectTrigger>
-                        <SelectContent>
-                          <SelectItem value="ASC">Ascending</SelectItem>
-                          <SelectItem value="DESC">Descending</SelectItem>
-                        </SelectContent>
-                      </Select>
-                    </div>
-                  </div>
-                  <Button
-                    size="icon"
-                    variant="ghost"
-                    onClick={() => onRemove(col)}
-                  >
-                    <Cross2Icon />
-                  </Button>
-                </div>
-              );
-            })}
-          </div>
-        </section>
-        <section className="flex flex-col">
-          <DropdownMenu>
-            <DropdownMenuTrigger asChild>
-              <Button variant="ghost" size="sm" className="flex justify-start">
-                <PlusIcon className="w-4 h-4 me-2 align-middle" /> Add sort
-              </Button>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent>
-              {fields.map((field) => (
-                <DropdownMenuItem
-                  key={field.name}
-                  onSelect={() => onAdd(field.name)}
-                >
-                  <FormFieldTypeIcon
-                    type={field.type}
-                    className="w-4 h-4 me-2 align-middle"
-                  />
-                  {field.name}
-                </DropdownMenuItem>
-              ))}
-            </DropdownMenuContent>
-          </DropdownMenu>
-          {isset && (
-            <PopoverClose asChild>
-              <Button
-                variant="ghost"
-                size="sm"
-                className="flex justify-start"
-                onClick={onReset}
-              >
-                <TrashIcon className="w-4 h-4 me-2 align-middle" /> Delete sort
-              </Button>
-            </PopoverClose>
-          )}
-        </section>
-      </PopoverContent>
-    </Popover>
-  );
-}
-
-function DataGridLocalSearch() {
-  const [state, dispatch] = useEditorState();
-
-  const onSearchChange = useDebounceCallback((txt: string) => {
-    dispatch({
-      type: "editor/data-grid/filter",
-      localsearch: txt,
-    });
-  }, 250);
-
-  return (
-    <Tooltip>
-      <TooltipTrigger>
-        <SearchInput
-          placeholder="Search locally"
-          onChange={(e) => onSearchChange(e.target.value)}
-          className="max-w-sm"
-        />
-      </TooltipTrigger>
-      <TooltipContent>Local search - Search within loaded data</TooltipContent>
-    </Tooltip>
-  );
-}
-
-function DataLoadingIndicator() {
-  const [state, dispatch] = useEditorState();
-  const { datagrid_isloading } = state;
-
-  const onRefresh = useCallback(() => {
-    dispatch({ type: "editor/data-grid/refresh" });
-  }, [dispatch]);
-
-  return (
-    <Button disabled={datagrid_isloading} onClick={onRefresh} variant="ghost">
-      <span className="me-2">
-        {datagrid_isloading ? <Spinner /> : <ReloadIcon />}
-      </span>
-      Refresh
-    </Button>
-  );
-}
-
 async function xsupabase_delete_query({
   form_id,
   main_table_id,
@@ -678,177 +428,8 @@ function table_keyword(
   }
 }
 
-function GridViewSettings() {
-  const [state, dispatch] = useEditorState();
-
-  const starwarsday = useMemo(
-    () => new Date(new Date().getFullYear(), 4, 4),
-    []
-  );
-
-  const tzoffset = useMemo(
-    () => s2Hmm(new Date().getTimezoneOffset() * -1 * 60),
-    []
-  );
-  const tzoffset_scheduling_tz = useMemo(
-    () =>
-      state.scheduling_tz
-        ? formatTZ(new Date(), "XXX", { timeZone: state.scheduling_tz })
-        : undefined,
-    [state.scheduling_tz]
-  );
-
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <button className="flex items-center justify-center">
-          <Badge variant="outline" className="cursor-pointer">
-            <GearIcon />
-          </Badge>
-        </button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="end" className="min-w-56">
-        <DropdownMenuLabel>Table Settings</DropdownMenuLabel>
-        <DropdownMenuSeparator />
-        <DropdownMenuCheckboxItem
-          checked={state.datagrid_filter.empty_data_hidden}
-          onCheckedChange={(checked) => {
-            dispatch({
-              type: "editor/data-grid/filter",
-              empty_data_hidden: checked,
-            });
-          }}
-        >
-          Hide records with empty data
-        </DropdownMenuCheckboxItem>
-        <DropdownMenuCheckboxItem
-          checked={state.datagrid_filter.masking_enabled}
-          onCheckedChange={(checked) => {
-            dispatch({
-              type: "editor/data-grid/filter",
-              masking_enabled: checked,
-            });
-          }}
-        >
-          Mask data
-        </DropdownMenuCheckboxItem>
-        <DropdownMenuSeparator />
-        <DropdownMenuRadioGroup
-          value={state.dateformat}
-          onValueChange={(value) => {
-            dispatch({
-              type: "editor/data-grid/dateformat",
-              dateformat: value as any,
-            });
-          }}
-        >
-          <DropdownMenuRadioItem value="date">
-            Date
-            <DropdownMenuShortcut>
-              {starwarsday.toLocaleDateString()}
-            </DropdownMenuShortcut>
-          </DropdownMenuRadioItem>
-          <DropdownMenuRadioItem value="time">
-            Time
-            <DropdownMenuShortcut>
-              {starwarsday.toLocaleTimeString()}
-            </DropdownMenuShortcut>
-          </DropdownMenuRadioItem>
-          <DropdownMenuRadioItem value="datetime">
-            Date Time
-            <DropdownMenuShortcut className="ms-4">
-              {starwarsday.toLocaleString()}
-            </DropdownMenuShortcut>
-          </DropdownMenuRadioItem>
-        </DropdownMenuRadioGroup>
-        <DropdownMenuSeparator />
-        <DropdownMenuRadioGroup
-          value={tztostr(state.datetz, "browser")}
-          onValueChange={(tz) => {
-            switch (tz) {
-              case "browser":
-                dispatch({ type: "editor/data-grid/tz", tz: LOCALTZ });
-                return;
-              default:
-                dispatch({ type: "editor/data-grid/tz", tz: tz });
-                return;
-            }
-          }}
-        >
-          <DropdownMenuRadioItem value="browser">
-            Local Time
-            <DropdownMenuShortcut>(UTC+{tzoffset})</DropdownMenuShortcut>
-          </DropdownMenuRadioItem>
-          <DropdownMenuRadioItem value="UTC">
-            UTC Time
-            <DropdownMenuShortcut>(UTC+0)</DropdownMenuShortcut>
-          </DropdownMenuRadioItem>
-          <DropdownMenuRadioItem
-            disabled={!state.scheduling_tz}
-            value={state.scheduling_tz ?? "N/A"}
-          >
-            Scheduling Time
-            {state.scheduling_tz && (
-              <DropdownMenuShortcut className="text-end">
-                {state.scheduling_tz}
-                <br />
-                (UTC{tzoffset_scheduling_tz})
-              </DropdownMenuShortcut>
-            )}
-          </DropdownMenuRadioItem>
-        </DropdownMenuRadioGroup>
-        <DropdownMenuSeparator />
-        <Link href={`./simulator`} target="_blank">
-          <DropdownMenuItem className="cursor-pointer">
-            <CommitIcon className="inline align-middle me-2" />
-            Open Simulator
-          </DropdownMenuItem>
-        </Link>
-      </DropdownMenuContent>
-    </DropdownMenu>
-  );
-}
-
-function s2Hmm(s: number) {
-  const now = new Date();
-  const startOfDayDate = startOfDay(now);
-  const updatedDate = addSeconds(startOfDayDate, s);
-  const formattedTime = format(updatedDate, "H:mm");
-
-  return formattedTime;
-}
-
 function txt_n_plural(n: number | undefined, singular: string) {
   return (n || 0) > 1 ? `${n} ${singular}s` : `${n} ${singular}`;
-}
-
-function MaxRowsSelect() {
-  const [state, dispatch] = useEditorState();
-
-  return (
-    <div>
-      <Select
-        value={state.datagrid_rows_per_page + ""}
-        onValueChange={(value) => {
-          dispatch({
-            type: "editor/data-grid/rows",
-            rows: parseInt(value),
-          });
-        }}
-      >
-        <SelectTrigger>
-          <SelectValue placeholder="rows" />
-        </SelectTrigger>
-        <SelectContent>
-          <></>
-          <SelectItem value={10 + ""}>10 rows</SelectItem>
-          <SelectItem value={100 + ""}>100 rows</SelectItem>
-          <SelectItem value={500 + ""}>500 rows</SelectItem>
-          <SelectItem value={1000 + ""}>1000 rows</SelectItem>
-        </SelectContent>
-      </Select>
-    </div>
-  );
 }
 
 function DeleteFieldConfirmDialog({
